+++ src/base/Link.php
@@ -182,9 +182,6 @@ abstract class Link extends SavableComponent implements LinkInterface
     {
         $rules = parent::rules();
         $rules[] = ['customLabel', 'string'];
-        // $rules[] = ['value', 'required', 'message' => Craft::t('linkit', 'A {type} is required.', [
-        //     'type' => 'value'
-        // ])];
         return $rules;
     }